more files moved; fixed imports; logs WIP
