+++ api/test/lib_test.py
@@ -5,7 +5,7 @@ Test oriented library
 import os
 from pathlib import Path
 
-from src.commons.json_utils import load_json  # pylint: disable=E0611
+from src.commons.utils import load_json  # pylint: disable=E0611
 
 
 def get_test_fixtures_path():
