+++ orchestrator/src/__init__.py
@@ -2,6 +2,14 @@
 Logging module based on common logger
 """
 
-from src.commons.logging import LoggerManager
+# import logging
 
-logger = LoggerManager(log_filepath="/opt/orchestrator/logs/worker.log").get_logger()
+# from src.commons.logging import InterceptHandler, LoggerManager
+# from worker import logger
+
+# # Also override Celery's logger
+# celery_logger = logging.getLogger("celery")
+# celery_logger.setLevel(logging.INFO)
+# celery_logger.addHandler(InterceptHandler())
+
+# logger = LoggerManager(log_filepath="/opt/orchestrator/logs/worker.log").get_logger()
